requirements.txt
# Core dependencies
python-telegram-bot==20.7
pymysql==1.1.0
python-dotenv==1.0.0

# AI and API
openai==1.3.5
aiohttp==3.9.1
beautifulsoup4==4.12.2
msal==1.25.0
requests==2.31.0

# PDF generation
reportlab==4.0.7

# Utilities
pytz==2023.3